requirements.txt
PyMuPDF>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
simplekml>=1.3.0
streamlit>=1.10.0
streamlit-drawable-canvas>=0.9.0
pandas>=1.3.0
numpy>=1.20.0
python-dateutil>=2.8.0
pytest>=6.0.0
matplotlib>=3.4.0

# MIDI SysEx functionality
mido>=1.2.10
python-rtmidi>=1.4.9
PyYAML>=6.0